Move and document deserializer module
