+++ cli/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tsh"
-version = "0.2.0"
+version = "0.3.1"
 authors = ["Sayan Nandan <ohsayan@outlook.com>"]
 edition = "2018"
