+++ corelib/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "corelib"
-version = "0.2.0"
+version = "0.3.1"
 authors = ["Sayan Nandan <ohsayan@outlook.com>"]
 edition = "2018"
